Refactor out into affinity.rs
